+++ R/gv_MRPP.R
@@ -19,6 +19,8 @@
 #'
 #' @param object (Required). a [`phyloseq::phyloseq-class`] or
 #' [`Biobase::ExpressionSet`] object.
+#' @param level (Optional). character. Summarization
+#' level (from \code{rank_names(pseq)}, default: NULL).
 #' @param variable (Required). character. grouping variable for test.
 #' @param method (Optional). character. Provide one of the currently supported
 #' options. See `distanceMethodList` for a detailed list of the supported options
@@ -43,11 +45,15 @@
 #' @importFrom Biobase pData exprs
 #' @importFrom stats setNames p.adjust
 #'
-#' @usage run_MRPP(object,
-#'                 variable,
-#'                 method,
-#'                 seedNum,
-#'                 alpha)
+#' @usage run_MRPP(
+#'    object,
+#'    level = c(NULL, "Kingdom", "Phylum", "Class",
+#'            "Order", "Family", "Genus",
+#'            "Species", "Strain", "unique"),
+#'    variable,
+#'    method = c("unifrac", "wunifrac", "GUniFrac", "bray", "dpcoa", "jsd"),
+#'    seedNum = 123,
+#'    alpha = 0.5)
 #'
 #' @export
 #'
@@ -59,15 +65,32 @@
 #' }
 #'
 run_MRPP <- function(
-              object,
-              variable,
-              method = "bray",
-              seedNum = 123,
-              alpha = 0.5) {
+    object,
+    level = NULL,
+    variable,
+    method = "bray",
+    seedNum = 123,
+    alpha = 0.5) {
+
+  # data("enterotypes_arumugam")
+  # object = enterotypes_arumugam
+  # level = "Phylum"
+  # variable = "Enterotype"
+  # method = "bray"
+  # seedNum = 123
+  # alpha = 0.5
 
   # phyloseq object
   if (all(!is.null(object), inherits(object, "phyloseq"))) {
     ps <- check_sample_names(object = object)
+
+    # taxa level
+    if (!is.null(level)) {
+      ps <- aggregate_taxa(x = ps, level = level)
+    } else {
+      ps <- ps
+    }
+
     if (!is.null(ps@phy_tree) & (method %in%
                                  c("unifrac", "wunifrac", "GUniFrac"))) {
       method <- match.arg(
@@ -80,7 +103,7 @@ run_MRPP <- function(
     }
     ## sample table & profile table
     sam_tab <- phyloseq::sample_data(ps) %>% data.frame() %>%
-      tibble::rownames_to_column("SampleID")
+      tibble::rownames_to_column("TempRowNames")
     if (phyloseq::taxa_are_rows(ps)) {
       prf_tab <- phyloseq::otu_table(phyloseq::t(ps)) %>%
         data.frame()
@@ -90,7 +113,7 @@ run_MRPP <- function(
   } else if (all(!is.null(object), inherits(object, "ExpressionSet"))) {
     # sample table & profile table
     sam_tab <- Biobase::pData(object) %>% data.frame() %>%
-      tibble::rownames_to_column("SampleID")
+      tibble::rownames_to_column("TempRowNames")
     prf_tab <- Biobase::exprs(object) %>% data.frame()
   }
   # distance